api: move UniformSamplingField from p3-challenger to p3-field
