+++ goldilocks/Cargo.toml
@@ -14,7 +14,6 @@ categories.workspace = true
 criterion.workspace = true
 
 [dependencies]
-p3-challenger.workspace = true
 p3-dft.workspace = true
 p3-field.workspace = true
 p3-mds.workspace = true
